pyproject.toml
[project]
name = "agents-experiments"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "openai-agents>=0.3.0",
    "ruff>=0.13.0",
    "trio>=0.31.0",
    "trio-asyncio>=0.15.0",
    "ty>=0.0.1a20",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]
